chore: move otelmw inside metrics
